requirements.txt
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
backcall==0.2.0
boto3==1.24.26
botocore==1.27.26
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
decorator==5.1.1py
defusedxml==0.7.1
Django==4.0.6
django-allauth==0.51.0
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
elastic-transport==8.1.2
elasticsearch==8.3.1
executing==0.8.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipython==8.4.0
jedi==0.18.1
jmespath==1.0.1
matplotlib-inline==0.1.3
msgpack==1.0.4
oauthlib==3.2.0
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.30
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.3.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
wcwidth==0.2.5
zope.interface==5.4.0
